requirements.txt
nltk==3.9.1
numpy==2.1.3
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.14.1
